+++ scripts/check_static_quality.sh
@@ -13,6 +13,7 @@ run_step() {
 run_step "gallery outputs" ./scripts/check_gallery_outputs.sh
 run_step "gallery metadata consistency" ./scripts/check_gallery_consistency.sh
 run_step "version metadata consistency" ./scripts/check_version_consistency.sh
+run_step "matlab-figure-ci dogfood version" ./scripts/check_mfigci_dogfood_version.sh
 run_step "roadmap status language" ./scripts/check_roadmap_status.sh
 run_step "maintainer activity bounds" ./scripts/check_maintainer_activity.sh
 run_step "citation metadata" ./scripts/check_citation.sh